backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
langchain>=0.3.28,<0.4
langchain-core>=0.3.50
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.3.0
sqlalchemy>=2.0.30
